refactor(orchestration): extract QuestionFlowCore as the third shared layer
Moves nineteen question-flow methods out of the facade verbatim (501 body
lines, confirmed byte-identical by script). 3970 -> 3428 lines. This unit
moves no instance state and needs no accessors: none of the nineteen calls
anything outside the nineteen.

The plan's QuestionFlowDeps Pick was wrong in both directions. It omitted
loadState and saveState, which four of the methods call, and listed config,
createId, logger and resumeWorkerTask, which appear zero times across all
nineteen bodies. The Pick is now exactly the five ports the code touches:
now, loadState, saveState, wakeCoordinatorSession, deliverCoordinatorMessage.
A unit constructible without a port it never calls is the point of the split.

buildCoordinatorRouteChatMetadata stays in the facade. The plan contradicts
itself here -- its Method Assignment lists the function among those that stay,
while the Task 10 section says to move it. The Method Assignment is right: its
sole caller is recordCoordinatorRouteContext, which belongs to Task 13, and
none of the nineteen reference it.

Four methods call kernel.mutate, which is non-reentrant. None was added,
removed, or relocated.
