+++ src/orchestration/orchestration-service.ts
@@ -20,8 +20,8 @@ import { AsyncMutex } from "./async-mutex";
 import { sameCoordinatorSession, stableCoordinatorSession } from "./coordinator-identity";
 import { sanitizeProgressSummary, stripProgressLines } from "./progress-line-parser";
 import { OrchestrationStateKernel } from "./service/orchestration-state-kernel";
+import { QuestionFlowCore } from "./service/question-flow-core";
 import { WorkerSessionManager } from "./service/worker-session-manager";
-import { isQuotaDeferredError } from "../weixin/messaging/quota-errors";
 import {
   DEFAULT_TASK_WATCH_POLL_INTERVAL_MS,
   DEFAULT_TASK_WATCH_TIMEOUT_MS,
@@ -127,7 +127,7 @@ export interface DeliverCoordinatorMessageRequest {
   text: string;
 }
 
-type FrozenCoordinatorDeliveryRoute = Pick<
+export type FrozenCoordinatorDeliveryRoute = Pick<
   DeliverCoordinatorMessageRequest,
   "chatKey" | "accountId" | "replyContextToken"
 >;
@@ -322,10 +322,12 @@ export interface OrchestrationGroupListFilter {
 export class OrchestrationService {
   private readonly kernel: OrchestrationStateKernel;
   private readonly workerSessions: WorkerSessionManager;
+  private readonly questionFlow: QuestionFlowCore;
 
   constructor(private readonly deps: OrchestrationServiceDeps) {
     this.kernel = new OrchestrationStateKernel({ logger: deps.logger }, deps.stateMutex);
     this.workerSessions = new WorkerSessionManager(deps, this.kernel);
+    this.questionFlow = new QuestionFlowCore(deps, this.kernel);
   }
 
 
@@ -1577,7 +1579,7 @@ export class OrchestrationService {
         });
       }
     } catch (error) {
-      await this.recordOpenQuestionWakeError(
+      await this.questionFlow.recordOpenQuestionWakeError(
         prepared.taskId,
         prepared.questionId,
         error instanceof Error ? error.message : String(error),
@@ -1609,19 +1611,19 @@ export class OrchestrationService {
         throw new Error(`task "${input.taskId}" does not exist`);
       }
 
-      this.assertCoordinatorOwnership(task, input.coordinatorSession);
+      this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
       if (task.status !== "blocked" && task.status !== "waiting_for_human") {
         throw new Error(`task "${input.taskId}" is ${task.status}, not blocked or waiting_for_human`);
       }
-      this.assertCoordinatorQuestionMatch(task, input.questionId);
-      this.assertTaskAnswerIsWithinAwaitedHumanSnapshot(state, task, input.questionId);
+      this.questionFlow.assertCoordinatorQuestionMatch(task, input.questionId);
+      this.questionFlow.assertTaskAnswerIsWithinAwaitedHumanSnapshot(state, task, input.questionId);
       if (!task.workerSession) {
         throw new Error(`task "${input.taskId}" does not have an assigned worker session`);
       }
 
       const now = this.deps.now().toISOString();
-      const packageRestore = this.captureTaskHumanPackageContext(state, task);
-      this.resolveTaskFromHumanPackage(state, task, now);
+      const packageRestore = this.questionFlow.captureTaskHumanPackageContext(state, task);
+      this.questionFlow.resolveTaskFromHumanPackage(state, task, now);
       task.status = "running";
       task.openQuestion = {
         ...task.openQuestion!,
@@ -1661,7 +1663,7 @@ export class OrchestrationService {
         answer,
       });
     } catch (error) {
-      await this.restoreBlockedQuestionAfterResumeFailure(
+      await this.questionFlow.restoreBlockedQuestionAfterResumeFailure(
         prepared.task.taskId,
         input.questionId,
         error instanceof Error ? error.message : String(error),
@@ -1671,7 +1673,7 @@ export class OrchestrationService {
     }
 
     if (prepared.closedPackageId) {
-      await this.handoffQueuedQuestions(prepared.task.coordinatorSession, prepared.closedPackageId);
+      await this.questionFlow.handoffQueuedQuestions(prepared.task.coordinatorSession, prepared.closedPackageId);
     }
 
     return prepared.task;
@@ -1689,7 +1691,7 @@ export class OrchestrationService {
         throw new Error(`task "${input.taskId}" does not exist`);
       }
 
-      this.assertCoordinatorOwnership(task, input.coordinatorSession);
+      this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
       const openQuestion = task.openQuestion;
       if (!openQuestion) {
         throw new Error(`task "${input.taskId}" does not have an open question`);
@@ -1797,16 +1799,16 @@ export class OrchestrationService {
         if (!task) {
           throw new Error(`task "${taskId}" does not exist`);
         }
-        this.assertCoordinatorOwnership(task, input.coordinatorSession);
+        this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
         if (task.status !== "blocked") {
           throw new Error(`task "${taskId}" is ${task.status}, not blocked`);
         }
-        this.assertCoordinatorQuestionMatch(task, questionId);
+        this.questionFlow.assertCoordinatorQuestionMatch(task, questionId);
         return task;
       });
 
       const now = this.deps.now().toISOString();
-      const route = this.snapshotCoordinatorDeliveryRoute(
+      const route = this.questionFlow.snapshotCoordinatorDeliveryRoute(
         this.kernel.ensureCoordinatorRoutes(state)[stableCoordinatorSession(input.coordinatorSession)],
       );
       if (coordinatorState.activePackageId) {
@@ -1858,7 +1860,7 @@ export class OrchestrationService {
               taskId: task.taskId,
               questionId: task.openQuestion!.questionId,
             })),
-            ...(route ? this.serializeFrozenDeliveryRoute(route) : {}),
+            ...(route ? this.questionFlow.serializeFrozenDeliveryRoute(route) : {}),
           },
         ],
       };
@@ -1898,7 +1900,7 @@ export class OrchestrationService {
       };
     }
 
-    await this.deliverHumanQuestionPackageMessage(prepared);
+    await this.questionFlow.deliverHumanQuestionPackageMessage(prepared);
     return {
       packageId: prepared.packageId,
       queuedTaskIds: prepared.queuedTaskIds,
@@ -1943,13 +1945,13 @@ export class OrchestrationService {
         throw new Error(`message "${input.messageId}" in package "${input.packageId}" is already delivered`);
       }
 
-      let route: FrozenCoordinatorDeliveryRoute | null = this.resolveFrozenPackageMessageRoute(message);
+      let route: FrozenCoordinatorDeliveryRoute | null = this.questionFlow.resolveFrozenPackageMessageRoute(message);
       if (!route) {
-        route = this.snapshotCoordinatorDeliveryRoute(
+        route = this.questionFlow.snapshotCoordinatorDeliveryRoute(
           this.kernel.ensureCoordinatorRoutes(state)[stableCoordinatorSession(input.coordinatorSession)],
         ) ?? null;
         if (route) {
-          Object.assign(message, this.serializeFrozenDeliveryRoute(route));
+          Object.assign(message, this.questionFlow.serializeFrozenDeliveryRoute(route));
         }
       }
 
@@ -1965,7 +1967,7 @@ export class OrchestrationService {
       };
     });
 
-    await this.deliverHumanQuestionPackageMessage(prepared);
+    await this.questionFlow.deliverHumanQuestionPackageMessage(prepared);
     return {
       packageId: prepared.packageId,
       messageId: prepared.messageId,
@@ -2008,7 +2010,7 @@ export class OrchestrationService {
       ) {
         return null;
       }
-      const messageTaskQuestions = this.resolveLiveMessageTaskQuestions(state, packageRecord, message);
+      const messageTaskQuestions = this.questionFlow.resolveLiveMessageTaskQuestions(state, packageRecord, message);
       if (messageTaskQuestions.length === 0) {
         return null;
       }
@@ -2054,7 +2056,7 @@ export class OrchestrationService {
     if (!activeMessage) {
       return null;
     }
-    const messageTaskQuestions = this.resolveLiveMessageTaskQuestions(state, packageRecord, activeMessage);
+    const messageTaskQuestions = this.questionFlow.resolveLiveMessageTaskQuestions(state, packageRecord, activeMessage);
 
     const openTaskQuestions = packageRecord.openTaskIds
       .map((taskId) => {
@@ -2111,7 +2113,7 @@ export class OrchestrationService {
       if (!task) {
         throw new Error(`task "${input.taskId}" does not exist`);
       }
-      this.assertCoordinatorOwnership(task, input.coordinatorSession);
+      this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
       if (!task.reviewPending) {
         throw new Error(`task "${input.taskId}" does not have a contested result`);
       }
@@ -2127,11 +2129,11 @@ export class OrchestrationService {
 
       if (input.decision === "discard") {
         replacementQuestionId = this.deps.createId();
-        const packageId = this.reopenActiveHumanPackageForTask(state, task, now);
+        const packageId = this.questionFlow.reopenActiveHumanPackageForTask(state, task, now);
         task.status = packageId ? "waiting_for_human" : "blocked";
         task.summary = "";
         task.resultText = "";
-        task.openQuestion = this.buildReplacementOpenQuestion(task, replacementQuestionId, now, packageId);
+        task.openQuestion = this.questionFlow.buildReplacementOpenQuestion(task, replacementQuestionId, now, packageId);
         this.kernel.appendTaskEvent(task, now, "attention_required", {
           status: task.status,
           message: task.openQuestion.question,
@@ -2169,7 +2171,7 @@ export class OrchestrationService {
           coordinatorSession: prepared.task.coordinatorSession,
         });
       } catch (error) {
-        await this.recordOpenQuestionWakeError(
+        await this.questionFlow.recordOpenQuestionWakeError(
           prepared.task.taskId,
           prepared.replacementQuestionId,
           error instanceof Error ? error.message : String(error),
@@ -2691,7 +2693,7 @@ export class OrchestrationService {
         return { task: { ...task }, shouldPropagate, closedPackageId: undefined as string | undefined };
       }
 
-      const closedPackageId = this.detachTaskFromQuestionFlows(state, task, now);
+      const closedPackageId = this.questionFlow.detachTaskFromQuestionFlows(state, task, now);
       const wasNeedsConfirmation = task.status === "needs_confirmation";
       task.status = "cancelled";
       if (wasNeedsConfirmation && task.summary.trim().length === 0) {
@@ -2721,7 +2723,7 @@ export class OrchestrationService {
       this.startWorkerCancellation(prepared.task);
     }
     if (prepared.closedPackageId) {
-      await this.handoffQueuedQuestions(prepared.task.coordinatorSession, prepared.closedPackageId);
+      await this.questionFlow.handoffQueuedQuestions(prepared.task.coordinatorSession, prepared.closedPackageId);
     }
 
     // I-2: non-running cancel transitions the task directly to `cancelled` without
@@ -2758,9 +2760,9 @@ export class OrchestrationService {
       let replacementQuestionId: string | undefined;
       if (task.correctionPending?.reason === "misrouted_answer") {
         replacementQuestionId = this.deps.createId();
-        const packageId = this.reopenActiveHumanPackageForTask(state, task, now);
+        const packageId = this.questionFlow.reopenActiveHumanPackageForTask(state, task, now);
         task.status = packageId ? "waiting_for_human" : "blocked";
-        task.openQuestion = this.buildReplacementOpenQuestion(task, replacementQuestionId, now, packageId);
+        task.openQuestion = this.questionFlow.buildReplacementOpenQuestion(task, replacementQuestionId, now, packageId);
         task.correctionPending = undefined;
         task.cancelRequestedAt = undefined;
         task.cancelCompletedAt = undefined;
@@ -2799,7 +2801,7 @@ export class OrchestrationService {
             coordinatorSession: prepared.task.coordinatorSession,
           });
         } catch (error) {
-          await this.recordOpenQuestionWakeError(
+          await this.questionFlow.recordOpenQuestionWakeError(
             prepared.task.taskId,
             prepared.replacementQuestionId,
             error instanceof Error ? error.message : String(error),
@@ -2869,7 +2871,7 @@ export class OrchestrationService {
     if (!currentTask) {
       throw new Error(`task "${input.taskId}" does not exist`);
     }
-    this.assertCoordinatorOwnership(currentTask, input.coordinatorSession);
+    this.questionFlow.assertCoordinatorOwnership(currentTask, input.coordinatorSession);
     this.assertNeedsConfirmation(currentTask);
 
     const workerSession =
@@ -2900,7 +2902,7 @@ export class OrchestrationService {
         if (!task) {
           throw new Error(`task "${input.taskId}" does not exist`);
         }
-        this.assertCoordinatorOwnership(task, input.coordinatorSession);
+        this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
         this.assertNeedsConfirmation(task);
         if (this.workerSessions.canStartParallelTask(state, task.targetAgent)) {
           return null; // capacity available — fall through to normal start
@@ -2951,7 +2953,7 @@ export class OrchestrationService {
         if (!task) {
           throw new Error(`task "${input.taskId}" does not exist`);
         }
-        this.assertCoordinatorOwnership(task, input.coordinatorSession);
+        this.questionFlow.assertCoordinatorOwnership(task, input.coordinatorSession);
         this.assertNeedsConfirmation(task);
         const previousStatus = task.status;
         const previousUpdatedAt = task.updatedAt;
@@ -3241,56 +3243,6 @@ export class OrchestrationService {
     }
   }
 
-  private assertCoordinatorQuestionMatch(task: OrchestrationTaskRecord, questionId: string): OrchestrationOpenQuestionRecord {
-    const openQuestion = task.openQuestion;
-    if (!openQuestion) {
-      throw new Error(`task "${task.taskId}" does not have an open question`);
-    }
-    if (openQuestion.questionId !== questionId) {
-      throw new Error(`task "${task.taskId}" open question is "${openQuestion.questionId}", not "${questionId}"`);
-    }
-    if (openQuestion.status !== "open") {
-      throw new Error(`task "${task.taskId}" question "${questionId}" is ${openQuestion.status}, not open`);
-    }
-    return openQuestion;
-  }
-
-  private assertTaskAnswerIsWithinAwaitedHumanSnapshot(
-    state: AppState,
-    task: OrchestrationTaskRecord,
-    questionId: string,
-  ): void {
-    if (task.status !== "waiting_for_human") {
-      return;
-    }
-
-    const packageId = task.openQuestion?.packageId;
-    if (!packageId) {
-      return;
-    }
-
-    const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-    if (coordinatorState.activePackageId !== packageId) {
-      return;
-    }
-
-    const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-    const awaitingMessageId = packageRecord?.awaitingReplyMessageId;
-    if (!packageRecord || !awaitingMessageId) {
-      return;
-    }
-
-    const awaitingMessage = packageRecord.messages.find((message) => message.messageId === awaitingMessageId);
-    const inSnapshot = awaitingMessage?.taskQuestions?.some(
-      (entry) => entry.taskId === task.taskId && entry.questionId === questionId,
-    );
-    if (!inSnapshot) {
-      throw new Error(
-        `task "${task.taskId}" question "${questionId}" is outside awaited message "${awaitingMessageId}" for package "${packageId}"`,
-      );
-    }
-  }
-
   private matchesFilter(task: OrchestrationTaskRecord, filter?: OrchestrationTaskFilter): boolean {
     if (!filter) {
       return true;
@@ -3305,14 +3257,6 @@ export class OrchestrationService {
       (filter.role === undefined || task.role === filter.role) &&
       (filter.status === undefined || task.status === filter.status)
     );
-  }
-
-  private assertCoordinatorOwnership(task: OrchestrationTaskRecord, coordinatorSession: string): void {
-    if (!sameCoordinatorSession(task.coordinatorSession, coordinatorSession)) {
-      throw new Error(
-        `task "${task.taskId}" belongs to coordinator "${task.coordinatorSession}", not "${coordinatorSession}"`,
-      );
-    }
   }
 
   private assertNeedsConfirmation(task: OrchestrationTaskRecord): void {
@@ -3378,492 +3322,6 @@ export class OrchestrationService {
     }
 
     return removedAny;
-  }
-
-  private snapshotCoordinatorDeliveryRoute(
-    route: OrchestrationCoordinatorRouteContextRecord | undefined,
-  ): FrozenCoordinatorDeliveryRoute | undefined {
-    if (!route) {
-      return undefined;
-    }
-
-    return {
-      chatKey: route.chatKey,
-      ...(route.accountId ? { accountId: route.accountId } : {}),
-      ...(route.replyContextToken ? { replyContextToken: route.replyContextToken } : {}),
-    };
-  }
-
-  private normalizeFrozenDeliveryRoute(route: FrozenCoordinatorDeliveryRoute): FrozenCoordinatorDeliveryRoute {
-    return {
-      chatKey: route.chatKey,
-      ...(route.accountId && route.replyContextToken
-        ? {
-            accountId: route.accountId,
-            replyContextToken: route.replyContextToken,
-          }
-        : {}),
-    };
-  }
-
-  private serializeFrozenDeliveryRoute(
-    route: FrozenCoordinatorDeliveryRoute,
-  ): Pick<
-    OrchestrationHumanQuestionPackageMessageRecord,
-    "routeChatKey" | "routeAccountId" | "routeReplyContextToken"
-  > {
-    const normalized = this.normalizeFrozenDeliveryRoute(route);
-    return {
-      routeChatKey: normalized.chatKey,
-      ...(normalized.accountId && normalized.replyContextToken
-        ? {
-            routeAccountId: normalized.accountId,
-            routeReplyContextToken: normalized.replyContextToken,
-          }
-        : {}),
-    };
-  }
-
-  private resolveFrozenPackageMessageRoute(
-    message: OrchestrationHumanQuestionPackageMessageRecord,
-  ): FrozenCoordinatorDeliveryRoute | null {
-    if (message.routeChatKey) {
-      return this.normalizeFrozenDeliveryRoute({
-        chatKey: message.routeChatKey,
-        ...(message.routeAccountId ? { accountId: message.routeAccountId } : {}),
-        ...(message.routeReplyContextToken ? { replyContextToken: message.routeReplyContextToken } : {}),
-      });
-    }
-
-    if (message.deliveredChatKey) {
-      return {
-        chatKey: message.deliveredChatKey,
-        ...(message.deliveryAccountId ? { accountId: message.deliveryAccountId } : {}),
-      };
-    }
-
-    return null;
-  }
-
-  private async deliverHumanQuestionPackageMessage(input: {
-    coordinatorSession: string;
-    packageId: string;
-    messageId: string;
-    promptText: string;
-    route: FrozenCoordinatorDeliveryRoute | null;
-  }): Promise<void> {
-    if (!input.route) {
-      const errorMessage = `coordinator "${input.coordinatorSession}" does not have a delivery route for human question packages`;
-      await this.recordPackageMessageDeliveryError(
-        input.coordinatorSession,
-        input.packageId,
-        input.messageId,
-        errorMessage,
-      );
-      throw new Error(errorMessage);
-    }
-    if (!this.deps.deliverCoordinatorMessage) {
-      const errorMessage = "deliverCoordinatorMessage dependency is required for human question package delivery";
-      await this.recordPackageMessageDeliveryError(
-        input.coordinatorSession,
-        input.packageId,
-        input.messageId,
-        errorMessage,
-      );
-      throw new Error(errorMessage);
-    }
-
-    try {
-      const deliveredRoute =
-        (await this.deps.deliverCoordinatorMessage({
-          coordinatorSession: input.coordinatorSession,
-          chatKey: input.route.chatKey,
-          accountId: input.route.accountId,
-          replyContextToken: input.route.replyContextToken,
-          text: input.promptText,
-        })) ?? input.route;
-      await this.recordPackageMessageDeliverySuccess({
-        coordinatorSession: input.coordinatorSession,
-        packageId: input.packageId,
-        messageId: input.messageId,
-        route: this.normalizeFrozenDeliveryRoute(deliveredRoute),
-        deliveryAccountId: deliveredRoute.accountId,
-      });
-    } catch (error) {
-      if (isQuotaDeferredError(error)) {
-        // Quota deferred is not a delivery failure: leave the package's
-        // delivery state pending so the next wake retries cleanly after the
-        // user's next inbound resets the quota window. Upstream callers
-        // (coordinatorRequestHumanInput, retryHumanQuestionPackageDelivery)
-        // will receive the deferred error
-        // and may need their own propagation handling — see follow-up TODO.
-        throw error;
-      }
-      await this.recordPackageMessageDeliveryError(
-        input.coordinatorSession,
-        input.packageId,
-        input.messageId,
-        error instanceof Error ? error.message : String(error),
-      );
-      throw error;
-    }
-  }
-
-  private async recordPackageMessageDeliverySuccess(input: {
-    coordinatorSession: string;
-    packageId: string;
-    messageId: string;
-    route: FrozenCoordinatorDeliveryRoute;
-    deliveryAccountId?: string;
-  }): Promise<void> {
-    await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, input.coordinatorSession);
-      if (coordinatorState.activePackageId !== input.packageId) {
-        return;
-      }
-
-      const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[input.packageId];
-      if (!packageRecord || packageRecord.status !== "active") {
-        return;
-      }
-
-      const message = packageRecord.messages.find((entry) => entry.messageId === input.messageId);
-      if (!message) {
-        return;
-      }
-
-      const now = this.deps.now().toISOString();
-      message.deliveredAt = now;
-      message.deliveredChatKey = input.route.chatKey;
-      message.deliveryAccountId = input.deliveryAccountId;
-      message.routeChatKey = input.route.chatKey;
-      message.routeAccountId = input.route.accountId && input.route.replyContextToken ? input.route.accountId : undefined;
-      message.routeReplyContextToken = input.route.replyContextToken;
-      message.lastDeliveryError = undefined;
-      packageRecord.awaitingReplyMessageId = input.messageId;
-      packageRecord.updatedAt = now;
-      await this.deps.saveState(state);
-    });
-  }
-
-  private async recordPackageMessageDeliveryError(
-    coordinatorSession: string,
-    packageId: string,
-    messageId: string,
-    errorMessage: string,
-  ): Promise<void> {
-    await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, coordinatorSession);
-      if (coordinatorState.activePackageId !== packageId) {
-        return;
-      }
-
-      const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-      if (!packageRecord || packageRecord.status !== "active") {
-        return;
-      }
-
-      const message = packageRecord.messages.find((entry) => entry.messageId === messageId);
-      if (!message) {
-        return;
-      }
-
-      const now = this.deps.now().toISOString();
-      message.lastDeliveryError = errorMessage;
-      if (packageRecord.messages.at(-1)?.messageId === messageId) {
-        packageRecord.awaitingReplyMessageId = undefined;
-      }
-      packageRecord.updatedAt = now;
-      await this.deps.saveState(state);
-    });
-  }
-
-  private async recordOpenQuestionWakeError(taskId: string, questionId: string, errorMessage: string): Promise<void> {
-    await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const task = state.orchestration.tasks[taskId];
-      if (!task || task.openQuestion?.questionId !== questionId || task.openQuestion.status !== "open") {
-        return;
-      }
-
-      task.openQuestion = {
-        ...task.openQuestion,
-        lastWakeError: errorMessage,
-      };
-      task.updatedAt = this.deps.now().toISOString();
-      await this.deps.saveState(state);
-    });
-  }
-
-  private async handoffQueuedQuestions(coordinatorSession: string, closedPackageId: string): Promise<void> {
-    const prepared = await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, coordinatorSession);
-      if (coordinatorState.activePackageId === closedPackageId) {
-        return { externalCoordinator: this.kernel.isExternalCoordinatorSession(state, coordinatorSession), queuedQuestions: [] };
-      }
-
-      const validQueuedQuestions = coordinatorState.queuedQuestions.filter((entry) => {
-        const task = state.orchestration.tasks[entry.taskId];
-        return (
-          task !== undefined &&
-          sameCoordinatorSession(task.coordinatorSession, coordinatorSession) &&
-          task.status === "blocked" &&
-          task.openQuestion?.status === "open" &&
-          task.openQuestion.questionId === entry.questionId
-        );
-      });
-      if (validQueuedQuestions.length !== coordinatorState.queuedQuestions.length) {
-        coordinatorState.queuedQuestions = validQueuedQuestions;
-        await this.deps.saveState(state);
-      }
-      return {
-        externalCoordinator: this.kernel.isExternalCoordinatorSession(state, coordinatorSession),
-        queuedQuestions: validQueuedQuestions,
-      };
-    });
-
-    if (prepared.queuedQuestions.length === 0 || prepared.externalCoordinator) {
-      return;
-    }
-
-    try {
-      await this.deps.wakeCoordinatorSession?.({
-        coordinatorSession,
-      });
-      await this.kernel.mutate(async () => {
-        const state = await this.deps.loadState();
-        const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, coordinatorSession);
-        coordinatorState.queuedQuestions = coordinatorState.queuedQuestions.filter(
-          (entry) =>
-            !prepared.queuedQuestions.some(
-              (queued) => queued.taskId === entry.taskId && queued.questionId === entry.questionId,
-            ),
-        );
-        await this.deps.saveState(state);
-      });
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : String(error);
-      await Promise.all(
-        prepared.queuedQuestions.map(async ({ taskId, questionId }) => {
-          const state = await this.deps.loadState();
-          const task = state.orchestration.tasks[taskId];
-          if (!task?.openQuestion || task.openQuestion.status !== "open" || task.openQuestion.questionId !== questionId) {
-            return;
-          }
-          await this.recordOpenQuestionWakeError(taskId, questionId, errorMessage);
-        }),
-      );
-    }
-  }
-
-  private async restoreBlockedQuestionAfterResumeFailure(
-    taskId: string,
-    questionId: string,
-    errorMessage: string,
-    packageRestore?: {
-      packageId: string;
-      packageRecord: OrchestrationHumanQuestionPackageRecord;
-      activePackageId?: string;
-    },
-  ): Promise<void> {
-    await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const task = state.orchestration.tasks[taskId];
-      if (!task || task.openQuestion?.questionId !== questionId || task.openQuestion.status !== "answered") {
-        return;
-      }
-
-      task.status = packageRestore ? "waiting_for_human" : "blocked";
-      task.openQuestion = {
-        ...task.openQuestion,
-        status: "open",
-        answeredAt: undefined,
-        answerSource: undefined,
-        answerText: undefined,
-        lastResumeError: errorMessage,
-      };
-      const now = this.deps.now().toISOString();
-      task.updatedAt = now;
-      if (packageRestore) {
-        const packageRecord = {
-          ...packageRestore.packageRecord,
-          initialTaskIds: [...packageRestore.packageRecord.initialTaskIds],
-          openTaskIds: [...packageRestore.packageRecord.openTaskIds],
-          resolvedTaskIds: [...packageRestore.packageRecord.resolvedTaskIds],
-          messages: packageRestore.packageRecord.messages.map((message) => ({ ...message })),
-          updatedAt: now,
-        };
-        this.kernel.ensureHumanQuestionPackages(state)[packageRestore.packageId] = packageRecord;
-        const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-        coordinatorState.activePackageId = packageRestore.activePackageId;
-      }
-      await this.deps.saveState(state);
-    });
-  }
-
-  private captureTaskHumanPackageContext(
-    state: AppState,
-    task: OrchestrationTaskRecord,
-  ):
-    | {
-        packageId: string;
-        packageRecord: OrchestrationHumanQuestionPackageRecord;
-        activePackageId?: string;
-      }
-    | undefined {
-    const packageId = task.openQuestion?.packageId;
-    if (!packageId) {
-      return undefined;
-    }
-
-    const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-    if (!packageRecord) {
-      return undefined;
-    }
-
-    const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-    return {
-      packageId,
-      packageRecord: {
-        ...packageRecord,
-        initialTaskIds: [...packageRecord.initialTaskIds],
-        openTaskIds: [...packageRecord.openTaskIds],
-        resolvedTaskIds: [...packageRecord.resolvedTaskIds],
-        messages: packageRecord.messages.map((message) => ({ ...message })),
-      },
-      activePackageId: coordinatorState.activePackageId,
-    };
-  }
-
-  private resolveTaskFromHumanPackage(state: AppState, task: OrchestrationTaskRecord, now: string): void {
-    const packageId = task.openQuestion?.packageId;
-    if (!packageId) {
-      return;
-    }
-
-    const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-    if (!packageRecord) {
-      return;
-    }
-
-    packageRecord.openTaskIds = packageRecord.openTaskIds.filter((taskId) => taskId !== task.taskId);
-    if (!packageRecord.resolvedTaskIds.includes(task.taskId)) {
-      packageRecord.resolvedTaskIds.push(task.taskId);
-    }
-    packageRecord.updatedAt = now;
-
-    const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-    if (packageRecord.openTaskIds.length === 0) {
-      packageRecord.status = "closed";
-      packageRecord.closedAt = now;
-      packageRecord.awaitingReplyMessageId = undefined;
-      if (coordinatorState.activePackageId === packageId) {
-        coordinatorState.activePackageId = undefined;
-      }
-    }
-  }
-
-  private detachTaskFromQuestionFlows(state: AppState, task: OrchestrationTaskRecord, now: string): string | undefined {
-    const questionId = task.openQuestion?.questionId;
-    const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-    if (questionId) {
-      coordinatorState.queuedQuestions = coordinatorState.queuedQuestions.filter(
-        (entry) => !(entry.taskId === task.taskId && entry.questionId === questionId),
-      );
-    }
-
-    const packageId = task.openQuestion?.packageId;
-    if (!packageId) {
-      return undefined;
-    }
-
-    const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-    if (!packageRecord) {
-      return undefined;
-    }
-
-    packageRecord.openTaskIds = packageRecord.openTaskIds.filter((taskId) => taskId !== task.taskId);
-    packageRecord.resolvedTaskIds = packageRecord.resolvedTaskIds.filter((taskId) => taskId !== task.taskId);
-    packageRecord.updatedAt = now;
-
-    if (packageRecord.openTaskIds.length === 0) {
-      packageRecord.status = "closed";
-      packageRecord.closedAt = now;
-      packageRecord.awaitingReplyMessageId = undefined;
-      if (coordinatorState.activePackageId === packageId) {
-        coordinatorState.activePackageId = undefined;
-      }
-      return packageId;
-    }
-
-    return undefined;
-  }
-
-  private reopenActiveHumanPackageForTask(
-    state: AppState,
-    task: OrchestrationTaskRecord,
-    now: string,
-  ): string | undefined {
-    const packageId = task.openQuestion?.packageId;
-    if (!packageId) {
-      return undefined;
-    }
-
-    const coordinatorState = this.kernel.ensureCoordinatorQuestionState(state, task.coordinatorSession);
-    if (coordinatorState.activePackageId !== packageId) {
-      return undefined;
-    }
-
-    const packageRecord = this.kernel.ensureHumanQuestionPackages(state)[packageId];
-    if (!packageRecord || packageRecord.status !== "active") {
-      return undefined;
-    }
-
-    if (!packageRecord.openTaskIds.includes(task.taskId)) {
-      packageRecord.openTaskIds.push(task.taskId);
-    }
-    packageRecord.resolvedTaskIds = packageRecord.resolvedTaskIds.filter((taskId) => taskId !== task.taskId);
-    packageRecord.updatedAt = now;
-    return packageId;
-  }
-
-  private buildReplacementOpenQuestion(
-    task: OrchestrationTaskRecord,
-    questionId: string,
-    askedAt: string,
-    packageId?: string,
-  ): OrchestrationOpenQuestionRecord {
-    const current = task.openQuestion;
-    return {
-      questionId,
-      question: current?.question ?? task.task,
-      whyBlocked: current?.whyBlocked ?? "Coordinator discarded the contested result",
-      whatIsNeeded: current?.whatIsNeeded ?? "A corrected answer from the worker",
-      askedAt,
-      status: "open",
-      ...(packageId ? { packageId } : {}),
-    };
-  }
-
-  private resolveLiveMessageTaskQuestions(
-    state: AppState,
-    packageRecord: OrchestrationHumanQuestionPackageRecord,
-    message: OrchestrationHumanQuestionPackageMessageRecord,
-  ): Array<{ taskId: string; questionId: string }> {
-    return (message.taskQuestions ?? [])
-      .filter((entry) => packageRecord.openTaskIds.includes(entry.taskId))
-      .filter((entry) => {
-        const task = state.orchestration.tasks[entry.taskId];
-        return (
-          task?.openQuestion !== undefined &&
-          task.openQuestion.status === "open" &&
-          task.openQuestion.questionId === entry.questionId
-        );
-      })
-      .map((entry) => ({ ...entry }));
   }
 
   private startWorkerCancellation(task: OrchestrationTaskRecord): void {
